Correct the waiting for "CONNECT" after doing CIPSTART
It still feels a bit clumsy the way it is done.

About waitForMessage(), we probably need to split up this function and
to have one that actually waits for the exact message. Right now it accepts
any line that _starts_ with the given message. This has bitten us with
the waiting for "CONNECT" which also accepted "CONNECT FAILED".