Wait 500 ms after switching off and "NORMAL POWER DOWN"
After this delay a much more reliable reading of the CTS pin is possible.
The 500ms is just a wild guess. Feel free to try shorter delays.

This update also changes first delay before the "power" toggle. The pin
should already be in a LOW state when we start toggle().